requirements.txt
annotated-types==0.7.0
anyio==4.3.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
cymem==2.0.8
# en-core-web-sm==3.7.1 (from https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl)
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jinja2==3.1.4
langcodes==3.4.0
language-data==1.2.0
marisa-trie==1.1.1
markupsafe==2.1.5
murmurhash==1.0.10
numpy==1.26.4
packaging==24.0
pip==24.0
preshed==3.0.9
pydantic==2.7.1
pydantic-core==2.18.2
python-telegram-bot==21.2
requests==2.32.2
setuptools==70.0.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.4
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
thinc==8.2.3
tqdm==4.66.4
typer==0.9.4
typing-extensions==4.12.0
urllib3==2.2.1
wasabi==1.1.2
weasel==0.3.4
wheel==0.43.0
